% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt}
\alias{adapt}
\title{tuning}
\usage{
adapt(theta, s, it)
}
\arguments{
\item{theta}{matrix, samples of posterior distribution. ncol: nb parameters, nrow: nb samples}

\item{s}{vector, proposal variances used to obtain posterior samples theta}

\item{it}{integer, number of theta sample}
}
\description{
tune the variance proposal
 try to tune variance toward 20% acceptance
}
\details{
return corrected variances to decrease/increase acceptance
}
