% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1_idx_inc.R
\name{f1_idx_inc}
\alias{f1_idx_inc}
\title{useful for indices}
\usage{
f1_idx_inc(x)
}
\arguments{
\item{x}{vector of 2 index integers}
}
\description{
useful for indices
}
\details{
return corrected variances to decrease/increase acceptance
}
