% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Like1.R
\name{Like1}
\alias{Like1}
\title{Compute the likelihood  for Poisson or NB offspring distributions}
\usage{
Like1(
  theta,
  data_long,
  t_window,
  n_loc,
  n_tw,
  param_agg = FALSE,
  overdispersion,
  p_reps
)
}
\arguments{
\item{data_long}{data.frame of incidence and overall infectivities by locations in long format 
(i.e. see  fct_MCMC_EpiEstim.R)}

\item{t_window}{integer, time window during which Rt is assumed constant.#'}

\item{n_loc}{number of locations}

\item{n_tw}{number of time windows (set in fct_MCMC_EpiEstim.R)}

\item{param_agg}{TRUE or FALSE if Rts estimates are aggregated by location}

\item{overdispersion}{TRUE or FALSE if overdispersion is assumed (Poisson vs. NB)}

\item{p_reps}{reporting probability, either a single real (if constant) or a vector 
with a value of reporting for each day. (set in fct_MCMC_EpiEstim.R)}

\item{theta0}{list of 2 vectors (for Rt and overdispersion) of inital values for parameters (set in MCMC_iter.R)}
}
\value{
logL log likelihood, a row vector of loglikelihoods. when param_agg is FALSE, 
logL has size number of independent time windows times number of location,
when param_agg is FALSE, logL has size number of independent time windows.
}
\description{
This is an internal function.
it compute the likelihood for given parameter set and the data.
Run the MCMC to sample the posterior of Rts (and optionally overdispersion) at each location.
this function is called internally by  MCMC_iter.R, adapt_tuning.R, and MCMC_Full.R, which is itself called in the 
fct_MCMC_EpiEstim.R.
}
