% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Like1.R
\name{Like1Poisson}
\alias{Like1Poisson}
\title{get likelihood  for Poisson or NB offspring distribution}
\usage{
Like1Poisson(theta, data_long, t_window, n_loc, n_tw, param_agg = FALSE)
}
\arguments{
\item{data_long}{dataframe of incidence and overall infectivities}

\item{t_window}{time windows}

\item{n_loc}{number of locations}

\item{n_tw}{number of time windows}

\item{param_agg}{TRUE or FALSE if Rt/overdispersion estimate are aggregated by location}

\item{theta:}{parameters values}
}
\description{
get likelihood  
internal to the MCMC
}
\details{
L log likelihood
}
