% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_iter.R
\name{MCMC_iter}
\alias{MCMC_iter}
\title{MCMC iterate}
\usage{
MCMC_iter(
  iter,
  theta0,
  s,
  data_long,
  n_loc,
  n_tw,
  t_window,
  prior,
  overdispersion,
  param_agg = FALSE
)
}
\arguments{
\item{iter}{integer, the number of iteration for the MCMC}

\item{theta0}{vector of inital parameters}

\item{s}{variance of proposal distributions (log-normal)}

\item{data_long}{dataframe of incidence and overall infectivities}

\item{n_loc}{number of locations}

\item{n_tw}{number of time windows}

\item{t_window}{time windows}

\item{prior}{prior for parameters}

\item{overdispersion}{TRUE or FALSE if overdispersion is assumed (Poisson vs. NB)}

\item{param_agg}{TRUE or FALSE if Rts estimates are aggregated by location}
}
\description{
run the MCMC to sample posterior of of Rts (and overdispersion) at each location
}
\details{
res a list containing matrices of: Rts posterior samples (before/after thinning), 
              overdispersions posterior samples (before/after thinning).
}
