% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_fct.R
\name{project_fct}
\alias{project_fct}
\title{project}
\usage{
project_fct(I0, Rt, n_sim, t_max, si, model = "poisson", size = NULL)
}
\arguments{
\item{theta}{matrix, samples of posterior distribution. ncol: nb parameters, nrow: nb samples}

\item{s}{vector, proposal variances used to obtain posterior samples theta}

\item{it}{integer, nro of theta}
}
\description{
wrapper for projection
}
\details{
lambda incidence weighted by serial interval
}
