% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Rt_sin.R
\name{sim_Rt_sin}
\alias{sim_Rt_sin}
\title{Simulate Rt}
\usage{
sim_Rt_sin(a, period, m, t_start, t_max)
}
\arguments{
\item{a}{amplitude of the wave}

\item{period}{period of Rts oscillations}

\item{m}{overall mean of Rt}

\item{t_start}{starting time for simulation of Rt}

\item{t_max}{end time}
}
\value{
Return data.frame with 2 columns for time (\code{t}) and the Rt (\code{$Rt})
}
\description{
simulate a sinusoidal Rt over time
}
\examples{

Rt <- sim_Rt_sin(a = 3, period = 20, m = 1, t_start = 1, t_max = 100)
plot(Rt$t, Rt$Rt, type='l')


}
