% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_incidence.R
\name{truncate_incidence}
\alias{truncate_incidence}
\title{truncate incidence}
\usage{
truncate_incidence(
  I0_t_import,
  I,
  t_window,
  t_truncate = NULL,
  incidence_truncate = NULL
)
}
\arguments{
\item{t_truncate}{integer, truncate incidence after t_truncate* t_window days}

\item{incidence_truncate}{integer, set incidence at NA for all time windows after which
'incidence_truncate' daily cases are reached}
}
\value{
A data.frame similar to the input incidence but with truncation applied.
}
\description{
Remove incidence (i.e. replace with NA) after a given time or once a certain incidence threshold is reached.
}
