% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_tuning.R
\name{adapt_tuning}
\alias{adapt_tuning}
\title{adaptive tuning}
\usage{
adapt_tuning(
  repli,
  within_iter,
  theta0,
  sigma,
  data_long,
  n_loc,
  n_tw,
  t_window,
  prior,
  overdispersion,
  param_agg = FALSE,
  p_reps
)
}
\arguments{
\item{repli}{number of time the variance of the proposal is tuned (10 tends to be ok)}

\item{within_iter}{iterations for evaluation of the acceptance with new proposal variances}

\item{sigma}{original variance to start with}

\item{others}{same as in MCMC_iter function}
}
\description{
tune proposal and give good initial values to start 'proper' MCMC
 try to tune variance toward 20% acceptance
}
\details{
res list of 2 vectors: theta0: parameters value (posterior samples) at the last iterations
                      sigma: new variances for the proposal distributions
}
