% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_EpiEstim.R
\name{fct_EpiEstim}
\alias{fct_EpiEstim}
\title{epiestim wrap}
\usage{
fct_EpiEstim(I0, I, t_window, mean_prior, std_prior, si)
}
\arguments{
\item{theta}{matrix, samples of posterior distribution. ncol: nb parameters, nrow: nb samples}

\item{s}{vector, proposal variances used to obtain posterior samples theta}

\item{it}{integer, nro of theta}
}
\description{
function to epiestim on the simulated data
}
\details{
lambda incidence weighted by serial interval
}
