% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_EpiEstim.R
\name{fct_EpiEstim}
\alias{fct_EpiEstim}
\title{EpiEstim wrapper}
\usage{
fct_EpiEstim(I0_t_import, I, t_window, mean_prior, std_prior, si, overlap)
}
\arguments{
\item{I0_t_import}{which of the initial incidence is imported}

\item{I}{dataframe of incidence. ncol: nb of location +1 (time), nrow: time}

\item{t_window}{integer, time window}

\item{mean_prior}{single real number, mean prior for Rts}

\item{std_prior}{single real number, stad deviation for prior for Rts}

\item{si}{serial distribution (as in EpiEstim include a 0 weighted SI on same day)}

\item{overlap}{TRUE/FALSE, whether using overlapping time window or not}
}
\value{
A list composed of 1 data.frame per location summarizing daily time, 
daily incidence (full incidence time series, including the imported cases for seeding), 
daily overall infectivity,the start and end of the time windows associated with 
the estimated mean/std deviation/median and
95% CI lower and upper quantile for Rts 
(i.e. Mean(R)", "Std(R)", "Quantile.0.025(R)", "Median(R)", "Quantile.0.975(R)").
If no overlapping windows is used, the entries for t_start/t_end and estimated Rt characteristics are only given 
when 't' is the end of the time window.
}
\description{
Wrapper to call EpiEstim independently on multiple locations.
}
\details{
Use EpiEstim independently on each location to estimate Rts without
 accounting for overdispersion, nor under-reporting.
}
