% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_MCMC_EpiEstim.R
\name{fct_MCMC_EpiEstim}
\alias{fct_MCMC_EpiEstim}
\title{wrapper for MCMC EpiEstim}
\usage{
fct_MCMC_EpiEstim(
  I0,
  I,
  t_window,
  mean_prior,
  std_prior,
  res_EpiEstim,
  overdispersion = FALSE,
  rep,
  thin = 10,
  param_agg = FALSE,
  Rt0_epiEstim = TRUE
)
}
\arguments{
\item{I0}{the incidence for the time window during which we assume Rt to be constant.
I is a dataframe, first column are dates then incidence for all locations
nb of row is the size of time widows, dates must be sequential}

\item{I}{integer of  numbers of locations}

\item{t_window}{integer, the number of iteration for the MCMC}

\item{mean_prior}{vector of inital parameters}

\item{std_prior}{variance of proposal distributions (log-normal)}

\item{res_EpiEstim}{Serial interval distribution (see SI_gamma_dist_EpiEstim)}

\item{rep}{number of time the variance of the proposal is tuned (10 tends to be ok)}

\item{thin}{iterations for evaluate the accpetance with new proposal variances}

\item{param_agg}{iterations for evaluate the accpetance with new proposal variances}

\item{Rt0_epiEstim}{iterations for evaluate the accpetance with new proposal variances}

\item{overdispersion:}{initial conidtions to guaranty that if R=1, then we predict the number of cases in the future will stablise at the mean number of cases observed in the time window
mu0 is also used as the mean of the (exponential) prior for intial conditions estimated}
}
\description{
run the MCMC to sample posterior of of Rts (and overdispersion) at each location, require
running first the 'fct_EpiEstim' (need to clean up input here so no duplications)
}
\details{
res a list containing 2 matrices: theta: matrix of posterior samples
                     and logL: matrix of associated log-likelihood
}
