% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_MCMC_EpiEstim.R
\name{fct_MCMC_EpiEstim}
\alias{fct_MCMC_EpiEstim}
\title{wrapper for MCMC EpiEstim}
\usage{
fct_MCMC_EpiEstim(
  I0_t_import,
  I,
  t_window,
  mean_prior,
  std_prior,
  res_EpiEstim,
  overdispersion = FALSE,
  rep,
  thin = 10,
  param_agg = FALSE,
  Rt0_epiEstim = TRUE,
  p_reps = 1,
  overlap = FALSE,
  input = NULL,
  mean_k_prior = 1000
)
}
\arguments{
\item{I0_t_import}{which of the initial incidence is imported}

\item{I}{I is a dataframe, first column are dates then incidence for all locations
nb of row is the size of time widows, dates must be sequential}

\item{t_window}{integer, time window}

\item{mean_prior}{mean prior for Rt}

\item{std_prior}{std deviation prior for Rt}

\item{res_EpiEstim}{starting at Rts at estimate from epiestim}

\item{rep}{iterations for MCMC, rep for the final sampling, rep/10 10 times for optimisation of proposal}

\item{thin}{thinning of posterior sample}

\item{param_agg}{iterations for evaluate the accpetance with new proposal variances}

\item{Rt0_epiEstim}{iterations for evaluate the accpetance with new proposal variances}

\item{overdispersion:}{FALSE/TRUE, TRUE: single overdispersion across locations}
}
\description{
run the MCMC to sample posterior of of Rts (and overdispersion) at each location, require
running first the 'fct_EpiEstim' (need to clean up input here so no duplications)
}
\details{
res a list containing 2 matrices: theta: matrix of posterior samples
                     and logL: matrix of associated log-likelihood
}
