% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Like1.R
\name{Like1}
\alias{Like1}
\title{get likelihood  for Poisson or NB offspring distribution}
\usage{
Like1(
  theta,
  data_long,
  t_window,
  n_loc,
  n_tw,
  param_agg = FALSE,
  overdispersion
)
}
\arguments{
\item{data_long}{dataframe of incidence and overall infectivities}

\item{t_window}{time windows}

\item{n_loc}{number of locations}

\item{n_tw}{number of time windows}

\item{param_agg}{TRUE or FALSE if Rts estimates are aggregated by location}

\item{overdispersion}{TRUE or FALSE if overdispersion is assumed (Poisson vs. NB)}

\item{theta:}{parameters values}
}
\description{
get likelihood  
internal to the MCMC
}
\details{
L log likelihood
}
