% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Like1.R
\name{Like1Poisson}
\alias{Like1Poisson}
\title{get likelihood}
\usage{
Like1Poisson(theta, data_long, t_window, n_sim, n_tw, param_agg = FALSE)
}
\arguments{
\item{lambda:}{'force of infection' matrix (incidence weighted by serial interval),
column number of days, row: number of locations}

\item{I}{matrix of observed incidence, same dimension as lambda}

\item{R0}{vector of reproduction numbers per locations}
}
\description{
get Poisson likelihood  
internal to the MCMC
}
\details{
L log likelihood
}
