% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_full.R
\name{MCMC_full}
\alias{MCMC_full}
\title{MCMC full}
\usage{
MCMC_full(
  iter,
  theta0,
  s,
  repli_adapt,
  within_iter,
  data_long,
  n_loc,
  n_tw,
  t_window,
  prior,
  overdispersion,
  thin,
  param_agg = FALSE
)
}
\arguments{
\item{iter}{integer, the number of iteration for the MCMC}

\item{theta0}{vector of inital parameters}

\item{s}{variance of proposal distributions (log-normal)}

\item{repli_adapt}{number of time the variance of the proposal is tuned (10 tends to be ok)}

\item{within_iter}{iterations for evaluation of the acceptance with new proposal variances, when tuning}

\item{data_long}{dataframe of incidence and overall infectivities}

\item{n_loc}{number of locations}

\item{n_tw}{number of time windows}

\item{t_window}{time windows}

\item{prior}{prior for parameters}

\item{overdispersion}{TRUE or FALSE if overdispersion is assumed (Poisson vs. NB)}

\item{thin}{nb of iterations being disregarded}

\item{param_agg}{TRUE or FALSE if Rt/overdispersion estimate are aggregated by location}
}
\description{
run the MCMC to sample posterior of Rts (and overdispersion) at each location + 
tuning of proposal variance + DIC of model
}
\details{
res a list containing matrices of: Rts posterior samples (before/after thinning), 
              overdispersions posterior samples (before/after thinning),
              likelihoods (before/after thinning), 
              DIC for the model.
}
