% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_fct.R
\name{project_fct}
\alias{project_fct}
\title{project}
\usage{
project_fct(I0, Rt, n_sim, t_max, si, model = "poisson", over = NULL)
}
\arguments{
\item{I0}{matrix of initial incidence per location (can be more than one day)}

\item{Rt}{dataframe with t (time) and associated Rts (instantaneous Rt)}

\item{t_max}{number of simulated time steps}

\item{si}{serial distribution (as in EpiEstim include a 0 weighted SI on same day)}

\item{model}{take 'poisson' or 'NB', i.e. offspring distribution assumed.}

\item{over}{value of overdispersion in the offspring distribution (here no under-reporting)}
}
\description{
wrapper for projection with/without overdispersion
}
\details{
I simulated incidence, inclusive of starting values
}
