% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_linear.R
\name{Rt_linear}
\alias{Rt_linear}
\title{project}
\usage{
Rt_linear(inp, n_week = 24)
}
\arguments{
\item{inp}{see others}
}
\description{
wrapper for projection with/without overdispersion
}
\details{
n_week nb of weeks for the simulation, default of 24
}
