% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_linear.R
\name{Rt_linear}
\alias{Rt_linear}
\title{Rt_linear}
\usage{
Rt_linear(inp, n_week = 24)
}
\arguments{
\item{inp}{input, need to be a data.frame and include \code{$R_max, $R_min,
$period, $step}: the max/min of the reproduction and the period of the 
pattern and duration of the step-changes (during 'step' days the Rt is constant).
Note that the ratio of period divided
by step should be an even integer to ensure that the Rt function is symmetrical.}

\item{n_week}{number of weeks for the simulation, default of 24 weeks}
}
\value{
Over the period of simulation (\code{n_week}), the function provide
Rt values following a periodic pattern with a number of step-wise changes, which
depends on the input.

A data.frame of time (\code{$t}) and daily Rt values (\code{$Rt}), 
size: (\code{n_week} x 7) x 2
}
\description{
Function to simulate a periodic and linearly changing Rt. the function allow 
some stepwise change of Rt.
}
\examples{
# with linear change in Rt (step = 1)
inp <- data.frame(R_max=.5, R_min=2,period = 16, step = 1)
Rt <- Rt_linear(inp = inp, n_week = inp$period)
plot(Rt$t, Rt$Rt, type='l')

# with linear step changes in Rt (step = 4)
inp <- data.frame(R_max=.5, R_min=2,period = 16, step = 4)
Rt <- Rt_linear(inp = inp, n_week = inp$period)
plot(Rt$t, Rt$Rt, type='l')

}
