% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Rt_sin.R
\name{sim_Rt_sin}
\alias{sim_Rt_sin}
\title{simulate}
\usage{
sim_Rt_sin(a, period, m, t_max, t_start)
}
\arguments{
\item{a}{matrix, samples of posterior distribution. ncol: nb parameters, nrow: nb samples}

\item{period}{vector, proposal variances used to obtain posterior samples theta}

\item{m}{integer, nro of theta}

\item{t_max}{vector, proposal variances used to obtain posterior samples theta}

\item{t_start}{integer, nro of theta}
}
\description{
simulate sinusoidal Rt
 try to tune variance toward 20% acceptance
}
\details{
lambda incidence weighted by serial interval
}
