% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Rt_sin.R
\name{sim_Rt_sin}
\alias{sim_Rt_sin}
\title{Simulate Rt}
\usage{
sim_Rt_sin(a, period, m, t_max, t_start)
}
\arguments{
\item{a}{amplitude of the wave}

\item{period}{period of Rts oscillations}

\item{m}{overall mean of Rt}

\item{t_max}{end time}

\item{t_start}{starting time for simulation of Rt}
}
\value{
Return data frame with 2 columns for time (\code{t}) and the Rt (\code{$Rt})
}
\description{
simulate a sinusoidal Rt over time
}
