% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_incidence.R
\name{truncate_incidence}
\alias{truncate_incidence}
\title{truncate incidence}
\usage{
truncate_incidence(
  I0_t_import,
  I,
  t_window,
  t_truncate = NULL,
  incidence_truncate = NULL
)
}
\arguments{
\item{I0_t_import}{integer, which of the initial incidence is imported.}

\item{I}{data.frame of incidence. ncol: number of locations +1 (\code{$time}), nrow: time}

\item{t_window}{integer, time window during which Rt is assumed constant.}

\item{t_truncate}{integer, truncate incidence after \code{t_truncate* t_window days + I0_t_import}.}

\item{incidence_truncate}{integer, set incidence at NA for all time windows after which
'incidence_truncate' daily cases are reached.}
}
\value{
A data.frame similar to the input incidence but with truncation applied.
}
\description{
Remove incidence (i.e. replace with NA) after a given time or once a certain incidence threshold is reached.
}
